# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: LabFlow API
  description: |
    LabFlow 实验室管理系统 API 文档
    - Base URL: http://localhost:8000
    - 认证方式: JWT Token (HttpOnly Cookie)
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development Server
paths:
...
  /api/v1/users/me:
    get:
      summary: 获取当前用户信息
      description: 获取已登录用户的信息
      tags:
        - User
      security:
        - CookieAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              example:
                code: 200
                message: success
                data:
                  id: 1
                  name: 张三
                  phone: "13800138000"
                  email: "user@example.com"
                  role: 2
                  is_active: 0
                  created_at: "2024-01-15T10:30:00Z"
        '401':
          description: 未登录
          content:
            application/json:
              example:
                code: 401
                message: 未登录
                data: null
    put:
      summary: 更新当前用户信息
      description: 更新已登录用户的信息（不包括密码和角色）
      tags:
        - User
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/UserResponse'
  /api/v1/users:
    get:
      summary: 获取用户列表
      description: 获取用户列表（管理员专用）
      tags:
        - User
      security:
        - CookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: 页码
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
          description: 每页数量
        - name: role
          in: query
          schema:
            type: integer
            enum: [0, 1, 2]
          description: 角色筛选 (0-超管, 1-管理员, 2-实验员)
        - name: is_active
          in: query
          schema:
            type: integer
            enum: [0, 1, 2]
          description: 状态筛选 (0-正常, 1-封禁, 2-注销)
        - name: keyword
          in: query
          schema:
            type: string
          description: 搜索关键词（姓名/手机/邮箱）
      responses:
        '200':
          description: 成功
          content:
            application/json:
              example:
                code: 200
                message: success
                data:
                  items:
                    - id: 1
                      name: 张三
                      phone: "13800138000"
                      email: "user@example.com"
                      role: 2
                      is_active: 0
                      created_at: "2024-01-15T10:30:00Z"
                  pagination:
                    page: 1
                    page_size: 20
                    total: 100
                    total_pages: 5
        '403':
          description: 权限不足
  /api/v1/users/{user_id}:
    get:
      summary: 获取用户详情
      description: 根据ID获取用户详情（管理员专用）
      tags:
        - User
      security:
        - CookieAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          description: 用户ID
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/UserResponse'
        '404':
          description: 用户不存在
    put:
      summary: 更新用户
      description: 更新用户信息（管理员专用，可修改角色和状态）
      tags:
        - User
      security:
        - CookieAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          description: 用户ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: 更新成功
        '403':
          description: 权限不足
        '404':
          description: 用户不存在
...
components:
  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          description: 用户ID
        name:
          type: string
          description: 用户名
        phone:
          type: string
          description: 手机号
        email:
          type: string
          description: 邮箱
        role:
          type: integer
          enum: [0, 1, 2]
          description: 角色 (0-超管, 1-管理员, 2-实验员)
        avatar_url:
          type: string
          description: 头像URL
        is_active:
          type: integer
          enum: [0, 1, 2]
          description: 状态 (0-正常, 1-封禁, 2-注销)
        last_login_at:
          type: string
          format: date-time
          description: 最后登录时间
        created_at:
          type: string
          format: date-time
          description: 创建时间
    UserUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 30
          description: 用户名
        email:
          type: string
          description: 邮箱
        avatar_url:
          type: string
          description: 头像URL
        role:
          type: integer
          enum: [0, 1, 2]
          description: 角色 (0-超管, 1-管理员, 2-实验员)
        is_active:
          type: integer
          enum: [0, 1, 2]
          description: 状态 (0-正常, 1-封禁, 2-注销)
...
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
...